od: Fixed progress display problem for od_processes.
Summary:
The progress information was not updated in the od_ProgressViewer, the reason being that the process-id(pid) was not passed to the progress viewer, which was then closing itself prematurely. 

So now I use a different overload of the QProcess::startDetached function to retrieve the process-id

Test Plan:
1) Luanched standared copy cube operation -> works OK
2) Created steering cube with cube from Petrel plugin -> Works OK

What does not work is the hiding of the balck console windwow. That means although you see the progress window in front but the od_process console window also stays there on screen. I did not find a convinient way from QProcess to hide this.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: bert, raman

Differential Revision: https://phabricator.dgbes.com:8443/D7736