Bugfix: unsigned 8-bit preloaded cubes did not show undef color.
Summary:
Bugfix: unsigned 8-bit preloaded cubes did not show undef color. See no reason
why this is not just an eight years old copy-and-paste error. May be this code
was never called if not doing preloading.

Test Plan:
Created unsigned 8-bit seismic cube in F3_Demo. Preloaded this volume, and
displayed it as an attribute on inline. Moving inline to 750 now shows the
color tables undef color instead of color 255. This is also confirmed in the
status bar info.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8122