od6.0 - added dragger move finished event in visRandomTrackDisplay
Summary:
This revision is same as the revision D7675 in od trunk.
The previous version, it always need to click mouse in the out side of survey box, then the seismic data can be updated after moving dragger position in Random line display.
Added this event which immediately updates seismic texture of random line.

Test Plan: Compiles and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8023