od6.0 - using seismic pre-load for RandomSeisDataPack
Summary:
This revision is same as revision D7674

The main issue for this implementation is to create RandomSeisDataPack based on cube datapack after pre-load. This revision added two functions setDataFrom()
in the class RandomSeisDataPack. The rests of changes in this revision making the thought clear and simple.

Test Plan: Compiles and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D8022