od5.0: Fix for displaying 2D data with non-zero Z0
Summary: I had a flattened survey where Z0 for both the survey and the 2D seismics started at -6000ms. When displayed, it was shifted downwards so as to start at 0. Hence this fix.

Test Plan: tested in flattened surveys for both 2D and 3D data and the display looks good. This code apparently does not apply to 3D.

Reviewers: helene

Reviewed By: helene

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D8162