od5.0: Fix for Crash while displaying Z slices
Summary: If you try to display a Z slide which is not at a sample position, OD crashes. Here, the array of the TypeSet was passed on to the ArrayValueSeries and eventually to the DataCube while the TypeSet itself was deleted being a local object.

Test Plan: Created a cube in F3 with 16ms sampling. Added default Z slice (924). No more crash. Data displays nicely.

Reviewers: nanne

Reviewed By: nanne

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D8214