Crash-fix at volume duplication (Mantis 4415).
Summary:
Crash-fix at volume duplication (Mantis 4415). The VolumeRenderScalarField
code was not made to change the shading mode after initialization. This is
what happened when switching off volume shading and next duplicate a volume
created earlier with shading on. An exceptional workflow, but it should not
crash. Additional to fixes D7997 and D8034.

Test Plan:
Switched off volume shading and next duplicated a volume created earlier with
shading on. No crash anymore. Also tested whether the added check does not
return directly when adding a new volume in the tree or when restoring a
volume.

Reviewers: nageswara

Reviewed By: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D8246