od: Changed the parenting mechanism from uiGraphicsItemGroup to uiGraphicsItem
Summary:
The whole mechanism of grouping uiGraphicsItems inside a BasemapObject has been changed to allow interaction with them.
Basically graphic items are now anchored to another graphic item that works as a parent and where all the view transformations are applied.

Test Plan:
HoverEvents are now working on BasemapObjects.
The default implementation is just an increase on the pen used to paint these items in order to give some sort of highlight effect.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D8270