Added support for setting the zoom-direction for the wheel.
Summary:
Per e-mail request. Default is changed. After discussion with support, it is suggested that this is settable somewhere, but where that is is not defined. Perhaps look and feel, or mouse interaction?

Perhaps mouse-interaction and look-and feel dialogs should be merged?

Test Plan: Works with both reversed and non-reversed setup.

Reviewers: nanne

Reviewed By: nanne

Maniphest Tasks: T253

Differential Revision: https://phabricator.dgbes.com:8443/D8295