Two bugfixes for SequentialReader (PORTING FROM od To od6.0 UNCHANGED)
Summary:
(1) sysMemCopy optimization in ArrayFiller (see D7262) no longer enabled
UndefArrayHandler::unsetUdf*(.) functions to adjust maximum amplitudes
that equal the undef value of the underlying data type. These were called
by Array3D<float>::set(.) before. New solution turns out only 25% slower
than single sysMemCopy and can do the undef handling for any underlying
data type.
(2) When being forced to run the original code in the else-branch of the
ArrayFiller::nextStep() function on my computer, new parallel tasks in
the SequentialReader queue are generated faster than being executed,
causing the reader to become slower and slower while being busy with
std::vector operations on a huge task queue. When finishing after a
long time, part of the data is missing and a crash occurs if pressing
abort before finishing. Apparently, something is running out of bound
if the length of the queue is not restricted.

Test Plan:
(1) Copied original seismics to a cube of unsigned chars. When preloading
this cube and passing it as mapped data to the random line visualisation
(new functionality to be committed soon), the pixels of maximum amplitude
are no longer showiing the undef color. Also checked that area without
valid data is still showing the undef color.
(2) SequentialReader runs the else-branch of the ArrayFiller::nextStep()
function without delays, crashes or missing data if the length of the
queue is restricted. Benchmarking did not show significant time differences
for maximum queue sizes of 100, 1000, and 1000. A queue size of only 10
turned out somewhat slower. My computer can run two processes in parallel.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D8301