od: Faster access for ArrayNDSlice
Summary: If the extracted ArrayNDSlice is the only available slice in the source ArrayND, then irrespective of the dimensions we are talking about, the memory is in order and we should exploit this situation for a faster access. For example, when I am accessing a Z slice in an Array3D and the only data available in that Array3D is that of the particular Z slice, the memory is in order.

Test Plan: Tested in 2D viewer. Display of Z slices is about 5 times faster (for pre-loaded data).

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D8330