Added third variable to Atmomic setIfValueIs.
Summary:
In many cases, you do not wish to have a reference as the first variable
of setIfValueIs, as you cannot throw in a hard-coded number. There are
however circumstances where you wish to know what the actual value is.

This update adds a pointer as a third variable where the actual value can
be put.

I considered making a new function, but it is very hard to find suitable names,
so I thought that using a default value was the best. All developers do however
need to catch this new usage (including 3rd party), and therfor the
default variable will only be activated in od7.

Test Plan:
Test program is updated, and seems to work. As this is an ifdef nightmare,
the true impact will only be seen through CDash.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D8319