od: Fix for multiple entries to .omf file while saving as a Petrel object
Summary:
Commit input was called several times and it kept on adding the object as a new objct as the name comparison failed for Petrel objects which has a (P) in their name.

So the input field is also updated while processing input with the new name so that the name comparison does not fail which prevents against multiple entries.

Test Plan: Added an od FaultStickSet to tree, saved as a Petrel FaultStickset, only one entry added to the .omf file

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D8365