od6.0: Enabled displaying gridline and annotation at special value.
Summary: Mantis: http://bugs.dgbes.com/view.php?id=2817. Removed epsaroundzero_ and made it to be determined from the range of the axis.

Test Plan: Tested in CCB and SynthRock and other 2D Viewers. Gridlines and annotation at start, stop and special value are always shown even if they collide with each other. Tested to see if only those gridlines that don't collide with existing gridlines are added after this.

Reviewers: bert

Reviewed By: bert

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D8379