od:  Change in the appearance of the wavelet display field in the Manage Wavelet dialog box
Summary: The wavelet display window is now top and bottom aligned to the multiselection box( selgrp object ). Label has been attached depicting the name of the displayed wavelet.

Test Plan: The change is working fine. The label is being updated everytime the wavelet is changed as well as the display is updated. In case of an empty wavelet there is no diplay in the display field. The alignment of the display field is maintained upon maximization.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D8448