od: Fix for Non Modality and GoOn msg of PreStack Event Import and Export Dialog Box
Summary: The fix now turns the import and export PreStack Event dialog box into non modal and will ask a que at the exnd of successful import/export for proceeding further or terminating the process

Test Plan:
The fix has beend tested and is working fine. The dialog boxes are closed automatically when
a. The OD window is closed
b. The survey is changed
The GoOn message is also popped up every time the action s successfully complete

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D8439