od6.0 - added undo, redo in planedatadisplay
Summary:
The requirement is asked by Paul and Marieke. After moving a plane data such as inline, we need undo and redo function to back to original position of inline.
The implementation is done by Ctrl+Z and Ctrl+Y

Test Plan: Compiles fine and Tests fine in inline, crossline and zslice

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8450