Clear eof-flag on new read-reqest.
Summary: If the input stream is a file that is currently written, or if it is a command that continiously outputs data, there may be more data available than at previous read. Hence, any attempt to read data should ignore the eof() state.

Test Plan: Test did occationally fail on windows before. Additional test was added to check that eof() was set after failed read to read past last byte.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D8578