od - added undo, redo in planedatadisplay
Summary:
This is the same revision as D8450 in od6.0
The requirement is asked by Paul and Marieke. After moving a plane data such as inline, we need undo and redo function to back to original position of inline.
The implementation is done by Ctrl+Z and Ctrl+Y

Test Plan: Compile and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8488