od - New class visEmStickSetDisplay
Summary:
This is the same revision as D8413 in od6.0
The new class is used for display knots( defined as markerset in this class ) in visFaultDisplay and visFaultStickSetDisplay. Perhaps the name is not good.
I have a little confuse, in general we call knot, in fault and faultstickset display, we call stick( see menu). So here I cal stickset too.

This class will be the inherited by visFaultdisplay and visFaultSticksetDisplay

Test Plan: Compiles fine and Test works in classes visFaultdisplay and visFaultSticksetDisplay.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8489