od- Added insertPos in visMarkset
Summary:
This is the same revision as D8381.
Although markerset has no relation with coordinates order. But if these coordinates are corresponding to a polyline,
then we need keep the coordinates order . So we need insertPos function

Test Plan: Compiles fine. Tests will be in od new feature of pick set editor.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8491