od: Centralized 2D line geometry handling during import.
Summary:
The same code was used at half a dozen places. So, this is an attempt to centralize it. It still depends on SeisTrcWriter to write the actual geometry. Later on we need to add a separate object to handle that.

Improved comment formatting

Test Plan: tested the single line case only by creating a 2D line from random line. Tested new/overwrite/non-overwrite all cases.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D8606