od6.0-handle Ctrl+a to z in visevent
Summary:
This revision is same as revision D8625.
The added code is based on the previous code in the function:remapKey() in od's QtKeyboardMap.
The reason is that remapKey() is not called during windows platform.

The current solution will handle both of linux and windows

Test Plan: Compiles and Tests fine

Reviewers: jaap

Reviewed By: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D8627