Fix for memory problem at callback.
Summary:
There was a problem with the detachCB function on the line 177 (see below). This lead that I had to change some other conditions in attachCB.

I am not 100% sure this is 100% ok, but this is one of the most used pieces of code, so I am sure we will know soon.

Test Plan: Existing test program works as expected without changes.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D8513