od: Convolution based interpolator
Summary:
This revision replaces the previous interpolation methods with a convolution (in the freq. domain) based interpolator
with a working similar to inverse distance algorithms. Two parameters determine the shape of the operator that itself determines the result after interpolation. The decay determines the weight of nearby points in the final result, while the taper angle determines whether interpolation is done only laterally (taperangle = 0 degrees) or also vertically (0 < taperangle < 90 ). When "smart scaling" is checked the true normalized results are shown. In the case of periodic data this leads to lower values in the middle since positive and negative values cancel each other out. When "smart" scaling is turned off, the samples are rescaled based on the RMS value of the trace they belong too.

This tool is meant for interpolating various types of 2D datasets such as seismmic, porosity, dip fields or wheelered seismic into 3D volumes. I plan to add a separate option for interpolating dip fields because the 2D dip inline and crossline components still have to be computed based on the direction of the repective 2D line. Another option I still want to add is "Horizon Cube - assisted"  interpolation. Also tests and warnings for the available amount of free memory still have to be added.

Test Plan: Compiles and tests fine on F3 Demo in the case of seismic and multicomponent 2D data on a Mac machine.

Reviewers: arnaud, kristofer, raman

Reviewed By: raman

Subscribers: dirkphilip, kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8529