od: Improved text for 2D data.
Summary:
For 2D data it is better to use 'attribute' in brackets so as to make it clear to the user what exactly it is, especially for those used to the Lineset era.
Also corrected an earlier typo.

Test Plan: The new name appears as expected in the I/O fields.

Reviewers: kristofer, bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D8653