[od6.0] Fixed bug 4968 - wrong index used on input data.
Test Plan: Tested through following the voxel connectivity workflow. Before the fix, the top of the array was written, not the corerct z-selection. After the fix, the correct output is written. The diff between the new output and the 5.0 output was computed for 1 line. It was null everywhere, apart from the stripes that were undefined (there were undefs in the data).

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D8665