od: Fix for missing SteeringCubes in the Manage Seismics window
Summary:
http://bugs.dgbes.com/view.php?id=4727. This was earlier reported by Chris Ross but we were not able to reproduce it. Now we can.
The reason was, the macro mIOObjContext returns a const reference to a static IOObjCtxt. In uiSeisSel this was altered using a const_cast. So, if you open a UI that uses 'NoSteering' constraint, the static IOObjCtxt of the SeisTrcTrnaslatorGroup gets modified and then when the same is used by Manage Seismics window you do not see any Steering Cubes.
Basically altering a const IOObjCtxt& using const_cast was not a good idea. Now this problem is gone. I think we also need to get rid of all the UI constructs using CtxtIOObj but that will be later.

Test Plan: Opened the Fault Likelihood window. Closed it. Opened the Manage Seismics window. All the cubes including the steering ones are still visible now. Before this fix the Steering cubes used to disappear after you open the Fault Likelihood window once.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D8697