Changed detachAllNotifiers to disconnectCallBacks() and included call to remove from main thread execution.
Summary:
With the new system for adding callbacks to be executed in the main thread,
there is a potential risk that something is deleted while it is waiting for the call from the main thread. There was then a function to avoid that,
CallBack::removeFromMainThread to avoid that, and a warning if it was not done.

As that would be two calls from the callback destructor, I thought that I should combine detachAllNotifiers and removeFromMainThread into one, which is done
in this revision.

On one hand, it may be an overkill to call removeFromMainThread from many
CallBack destructors, while it is only sometime needed. The call is
however quite quick, so I don't expect any penalyt.

Test Plan: New tests are added to ensure that this works (committed in earlier revisions).

Reviewers: nanne

Reviewed By: nanne

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D8713