od: Add icon to attribute right-click menu and uiAttrSel to identify pre-loaded data
Summary: Request from BG to show which data has been preloaded

Test Plan:
Preload data and right-click on treeitem shows the icon in the submenu.
Add and select data pops up a selection dialog, also there the icon is in front of the preloaded data.
Also works for steering cubes

Reviewers: helene

Reviewed By: helene

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D8739