od: Fixed a regression
Summary: Just realized that CtxtIOObj are not really the nuisance I thought them to be. They hold the current IOObj pointer and hence cannot be done away with. Therefore, Undid some of my changes from last week and made an alternative fix for the const CtxtIOObj issue.

Test Plan: Tested several uiIOObjSelDlg instances, the current selection is preserved in the dialog.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D8754