Bugfix: CmdRecorder was writing empty name for OpendTect Main Window (reported by Lakshmi)
Summary:
CmdRecorder was writing empty name for OpendTect Main Window. Although in
principle CmdDriver can deal with this, it is highly confusing to the user
and ruins portability. Empty name was recently introduced by "program name"
revision of uiodmain.cc (see Phabricator D8349) without coercive reason.
If such a reason appears, a makeshift in the CmdDriver plugin would be an
alternative.

Test Plan:
CmdRecorder is writing [OpendTect Main Window] again, which seems to be used
only internally and is not displayed as title on the top window bar.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8769