od6.0-od - improvement on calling setSelectionMode() for fixing 4693 and 4757
Summary:
The bugs are not completely solved yet. At least this is one of steps. In order to avoid more works caused by API freezing.
I would like first do this revision.

Test Plan:
Compiles and Tests as better thane previouis od version when there are multi objects in the secne such as Fault plane,
Fault stickset and pickset, polygon and so on.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D8809