Fix for Examine file in 'Table' mode.
Summary:
http://bugs.dgbes.com/view.php?id=4760.
The function putRow was returning a const char* which has now changed to a uiString. For the const char* value 3 cases were handled: null string, empty string and non-empty string. Now uiString gives jsut one state for the first two and this leads to an infinite loop in uiFileEdit.

The return type has now been changed to bool and the uiString& argument still gives two cases: empty and non-empty.

This class is too complicated to be used by the plugin developers. So, can we port it to 6.0 post today's API freeze?

Test Plan: Examine buttons work for Horizon, Well import etc...

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D8827