Fix for bug 4771
Summary:
There was a nasty bug in od, od6 and od6 that makes the surface limited filler always use auxdata 0 instead of the real one.

This patch looks up what name the data has, and finds the correct index for the
auxdata.

Test Plan: Tested by using two aux data inputs, both for velocity and for gradient. Results seem to be good.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D8839