od-make display intersection point between horizon and inline,cross line and z-slice working
Summary:
The current od versions doesn't work on display of the intersection single point between a horizon and in/cr lines, because current od
1. uses markerseeds display both of emobject seeds and intersection node, in function:addDataToMarker() of emhorizonpainter3d.cc,
we display only seeds with the condition "if ( hor3d.isPosAttrib(posid,EM::EMObject::sSeedNode())", if we disable this line, the intersection node can
be displayed. But the problem is when we display a tracking horizon line on the section, all seed nodes are displayed.
2. The function enableSeeds(bool) is called when loaded a existing horizon by Vw2DHorizon2D::draw(), during the calling, if horizon is not in the 
tracking mode, seeds will be not set as display. thus, intersection node is not displayed too.

To solve above problem, I prefer to introduce a temporal node called intersection node in emobject node definition. Thus we can distinguish seed node 
and intersection node. The intersection node is used only for displaying, it doesn't influence any properties of emobjects. If we have intersection node,
The displayed node number is 1.
At the same time, this revision makes intersection node display property same as polyline. 
The seed node properties are same as emobject seed node propties.

Test Plan: Compiles fine. Tests fine with F3Demo from 3d to 2d, and in 2d insert in line and crosslines

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8856