[od6.0] Rework of the enum tools.
Summary:
The enums have been revistied with the following objectives:

1. Enable localizations
2. Simplify (i.e. remove macros).

1 & 2 are achieved by moving function implementation into a new template
class: EnumDefImpl, where all parsing is done. The functions which are
still in the macros are only for legacy. The plan is that these should be
marked as depricated in od, and be removed - perhaps in od version 8.
The idea is that everything should be done through the static function Def(),
which returns the EnumDefImpl class.

2: The combo-boxes and StringListImpDef are made accept the EnumDef class, and
thus getting the translated text.

The idea is that the translated text is ONLY for viewing, not for selection. This makes it possible to change the uiString (i.e. the viewing) without
changing the definition (i.e. par-files). This may however lead to problems
when an uiComboBox is checked for its text() and not the id or the index. Initially, this is not a problem as the uiStrings now are auto-generated from the
names, but as more uiStrings that do not conform with the name are added, the problem may grow. One option to consider is to make uiComboBox never return the text of the selected item (and same with uiGenInput with a combo-box).

Test Plan: OpendTect starts correctly comboboxes still work.

Reviewers: nanne

Reviewed By: nanne

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D8857