[od6] Made label clearer on what data is expected.
Summary:
The gradient data is expected to be in SI units when reading it from the surface data, but it is in /ms when entered as a constant.
This is confusing to the user, and while we may change the policy later on, the UI should make it clear what is expected.

Test Plan: New label comes up nicely.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D8887