Crashfix: Cloning of faultstickknots during dragging crashes ocasionally.
Summary:
(1) Cloning of faultstickknots during dragging crashes ocasionally. This is a
relatively unknown feature. The user can clone a yellow knot of a faultstick
by pressing a specific key while start dragging it. Unlike Coin3D, it turns
out that OSG crashes when inserting a new knot during the motion callback of
the dragger. Without really understanding the precise cause of the crash, it
no longer happens when using a timer to delay the knot insertion till after
the callback finishes.
(2) Changed the specific key to clone a faultstick knot from Space-key to
Backquote-key, because the Space-key is going to be used in switching from
pick mode to drag mode and vice versa.

Test Plan:
(1) Cloned a whole lot of faultstick knots without crashing, while it happened
quite regularly before. 
(2) Faultstick knots are indeed cloned by pressing Backquote instead of Space.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8812