Bugfix: sticks of FaultStickSet still editable after edit button turned off (Mantis 4672). (PORTING FROM od TO od5.0 WITH !viseditor_->isOn() CHECK)
Summary:
Bugfix: sticks of FaultStickSet were editable after edit button turned off
(Mantis 4672). The new "updateManipulator()" solution is completely adopted
from FaultDisplay, where the same functionality was already working fine.
Note that this bug was previously fixed by Ding in od and od6.0, but
apparently not ported to od5.0, so it does not need to be reverted here.

Test Plan: Sticks no longer editable when the edit is button off.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8908