[od6] Enabled settings that slows down mouse wheel movement on mac, as the trackpad is more sensitive than a wheel.
Summary:
http://bugs.dgbes.com/view.php?id=4775 says that the trackpad is extremely sensititive with regards to zooming in and out.

This fixes that by setting the zoom speed through a user setting.

It was the original idea that the zoom-speed would be translated into the 2D
viewers as well, but it turned out to be hard. The longer term solution should be to rebuild all wheel-zooming to not emulate zoom-button-pushes, but go in and zoom with a variable zoom-amount at every call.

Test Plan: Tested with both trackpad and mouse on a mac, and works as expected.

Reviewers: mahant

Reviewed By: mahant

Differential Revision: https://phabricator.dgbes.com:8443/D8929