Implemented spacebar toggle from picking mode to position mode and back. (PORTING FROM od TO od6.0 UNCHANGED)
Summary:
(1) Implemented spacebar toggle from picking mode to position mode and back.
While in act mode, it enables the user to toggle between picking seeds and 
moving planar/volume draggers without having to select the dragging/picking
object in the tree and losing eye focus. If more than one dragger is available
in the scene, the one (most recently) under the mouse pointer is selected.
(2) Prevented that visBase::EventCatchHandler passes keyboard events if the
mouse pointer is not in the scene but the scene still has keyboard focus.
(3) Solved interference with rarely used faultstick knot cloning, which also
occupied the spacebar, but will now occupy the backqoute-key instead.

Test Plan:
Played around with one or more position mode manipulator objects and one or
more picking objects in one and also in two scenes. Everything seems to work
consistently within the preconditions of the functionality requested by the
users.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8879