[od6] Refined the methods of EnumDefs.
Summary:
I wanted the enum functions to be crystal clear in what they do. This goes in three steps, of
which this is the first:

Here, all new methods are laid out and introduces.
Next revision, which is much larger, make all OD code use the new functions instead of the old functions, and the
thrird revision will mark the old code as deprecated and lock it down.

Test Plan: Sanity check is done.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D8960