od : Synchornisation for 2D Viewers & Basemap
Summary:
It is a preliminary revision for synchronisation. If the basic concept is approved, I would incorporate other data types to it. For now I have synchronised only the Inl/Crl & Horizon3D.

uiViewer2D has a Viewer2DDataMan which keeps account of the number of Viewer2DDataObjects that are added to the 2D Viewer. So I have tied the BaseMap SyncObject to the Viewer2DDataObject. A Viewer2DDataObject is created when a new item is added to the tree. But a single Horizon can be present in several 2D Viewers. So I have changed the tied a TypeSet of Viewer2DDataObject ids to a particular SyncObject.

Similar to the 3D when a new Viewer2DDataObject is added to the Viewer2D a callback will be catched by the uiBaseMapSync which will create s new SyncObject.

I have made the ids for Viewer2DDataObject unqiue so that we can use it properly. The vw2did_ is the viewer 2d.

Test Plan: Sync is working properly , both ways. However I need some input regarding what to do with Show in 2D Viewer options. Because for that we need to do some more changes

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D8933