od6.0: Possible fix for crash in PicksetDisplay while changing the survey
Summary:
http://bugs.dgbes.com/view.php?id=4856

The callback order was the problem. The Picksets were removed first in the callback while they were still displayed.

So made the "void uiODApplMgr::surveyToBeChanged()" as the fist callback, which will remove the display first later the PiackSet will be removed from the manager.

Test Plan: Tested with all the objects that were using IOM().surveyToBeChanged notifier. and did not get any crash so far

Reviewers: nanne, raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9002