od: Implemented the OD::StdIconType
Summary:
To make StdIconType correspond to the uiString::sKeys, I wanted to have the same enum member. Turns out a lot of those were already in the OD::KeyBoardKey.
This enum has many very general names, like 'Cancel' A-Z ... Help ...
So I added a prefix to that one.
There is also a dGB part, not worth review as it's just KB_ prefixing only.

Test Plan: Not a lot to test, seems to work.

Reviewers: nanne, kristofer

Reviewed By: kristofer

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D9054