od6.0: Set a correct zrg
Summary:
This seems coming back again and again.
If you load a volume builder chain on a z-slice, it ends up computing the whole volume for the chain. Not sure in what case we should ignore the steoutzrg.

Test Plan: Tested on F3 with on the fly loading Volume Processing as an attribute, works fine.

Reviewers: nanne, raman

Differential Revision: https://phabricator.dgbes.com:8443/D9044