od6.0-changes in horizon tracking between seeds
Summary:
Concerning the name and mode in GUI still is discussing.
This revision only make the operation as same as manual tracking except from the difference in snap position.
After investigation, actually it is simple to reach the goal in implementation because we have done a new manual tracking mode.

Test Plan: Compiles fine. Tests in 2d, 3d horizon in 3dviewer and 2dviewer working fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9106