od6.0: Dragger released was called just by releasing mouse button.
Summary:
I have added a bool to trigger the release only if there is any movement. In connection to bug report

http://bugs.dgbes.com/view.php?id=4947

what is happenning here is that, with every click the dragger release was being triggered which in essence triggereing the update display with the uiTaskRunner.
The uiTaskrunner was stealing the mouse focus from the 3D window, thus page/pagedown was not affecting the scene.

Test Plan:
1) Added inline/cr/zslice, and dragged -> works OK
2) Tested with Random Line drag movements even with pre-loaded data -> works OK
3) Tested with pageup/down with multicomponent data -> works OK

Reviewers: jaap

Reviewed By: jaap

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9123