Fix for http://bugs.dgbes.com/view.php?id=4981
Summary:
Inserted a notifier to notify changes in the random line name.

Note: It was easier, for now, to declare de nameChanged notifier in the
RandomLine class than in the NameObject one since the latter would need changes
in its deleteNofity system.

Test Plan: It should keep the same behaviour as before.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D9188